~uv test uses ts-node
