+++ test/uv/.gitignore
@@ -1,0 +1,2 @@
+dump.json
+syntax.json
