+++ test/uv/run.ts
@@ -1,0 +1,2 @@
+import './build';
+import './parse';
